+++ kernel/Cargo.toml
@@ -27,7 +27,7 @@ fastrand.workspace = true
 abort.workspace = true
 panic-unwind2.workspace = true
 util.workspace = true
-async-exec.workspace = true
+kasync = { workspace = true, features = ["panic-unwind2", "counters"] }
 
 # 3rd-party dependencies
 rustc-demangle.workspace = true
