+++ kernel/src/arch/riscv64/trap_handler.rs
@@ -298,13 +298,9 @@ extern "C-unwind" fn default_trap_handler(
                 }
             }
             Trap::Interrupt(Interrupt::SupervisorTimer) => {
-                if let Some((expired, Some(next_deadline))) = global().executor.timer().try_turn() {
+                if let Some((_, Some(next_deadline))) = global().executor.timer().try_turn() {
                     // Timer interrupts are always IPIs used for sleeping
                     sbi::time::set_timer(next_deadline.ticks.0).unwrap();
-
-                    if expired > 0 {
-                        global().executor.unpark_one();
-                    }
                 } else {
                     // Timer interrupts are always IPIs used for sleeping
                     sbi::time::set_timer(u64::MAX).unwrap();
