+++ kernel/src/main.rs
@@ -50,12 +50,12 @@ use crate::mem::bootstrap_alloc::BootstrapAllocator;
 use crate::state::{CpuLocal, Global};
 use abort::abort;
 use arrayvec::ArrayVec;
-use async_exec::executor::{Executor, Worker};
-use async_exec::time::{Instant, Ticks};
 use cfg_if::cfg_if;
 use core::range::Range;
 use core::slice;
 use fastrand::FastRand;
+use kasync::executor::{Executor, Worker};
+use kasync::time::{Instant, Ticks};
 use loader_api::{BootInfo, LoaderConfig, MemoryRegionKind};
 use mem::PhysicalAddress;
 use mem::frame_alloc;
