+++ kernel/src/shell.rs
@@ -20,13 +20,13 @@ use crate::mem::{Mmap, PhysicalAddress, with_kernel_aspace};
 use crate::state::global;
 use crate::{arch, irq};
 use alloc::string::{String, ToString};
-use async_exec::executor::Executor;
 use core::fmt;
 use core::fmt::Write;
 use core::ops::DerefMut;
 use core::range::Range;
 use core::str::FromStr;
 use fallible_iterator::FallibleIterator;
+use kasync::executor::Executor;
 use spin::{Barrier, OnceLock};
 
 static COMMANDS: &[Command] = &[PANIC, FAULT, VERSION, SHUTDOWN];
